+++ external/anomaly/configs/base/configuration.py
@@ -19,7 +19,7 @@ Configurable parameters for anomaly classification task
 from sys import maxsize
 
 from attr import attrs
-from configs.utils.configuration_enums import ModelBackbone, POTQuantizationPreset
+from configs.base.configuration_enums import ModelBackbone, POTQuantizationPreset
 from ote_sdk.configuration import ConfigurableParameters
 from ote_sdk.configuration.elements import (
     ParameterGroup,
